backend/pyproject.toml
[project]
name = "sefs-backend"
version = "0.1.0"
description = "Semantic Entropy File System — Backend Engine"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "watchdog>=4.0.0",
    "PyMuPDF>=1.24.0",
    "python-docx>=1.1.0",
    "markdown>=3.6",
    "pandas>=2.2.0",
    "openai>=1.30.0",
    "ollama>=0.3.0",
    "hdbscan>=0.8.40",
    "umap-learn>=0.5.7",
    "numpy>=1.26.0",
    "scikit-learn>=1.5.0",
    "numba>=0.60.0",
    "llvmlite>=0.43.0",
    "pynndescent>=0.5.13",
    "aiosqlite>=0.20.0",
    "websockets>=12.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.7.0",
    "chardet>=5.2.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
